* debian/init:   - Remove already ran detection   - Log the running of ec2-run-user-data to /var/log/ec2-user-data.log * ec2-set-hostname.py:   - set hostname to the Ec2 local-hostname   - Update the /etc/hosts to change the ubuntu hostname to the     public hostname. * ec2-fetch-credentials:   - Copy the ssh keys to the ubuntu user.   - Setup authorized keys for root to tell the user to login as the     ubuntu user when they try to connect. * ec2-run-user-data:   - Create an .already-ran file to check to see if ec2-run-user-data     already ran.   - Save the ec2-run-user-data script in /var/ec2.

+++ ec2-run-user-data.py
@@ -36,21 +36,27 @@ def get_user_data():
     return data
 
 def get_ami_id():
-    url = 'http://169.254.169.254/%s/meta-data', % api_ver
+    url = 'http://169.254.169.254/%s/meta-data' % api_ver
     ami_id = urllib.urlopen('%s/ami-id/' %url).read()
     return ami_id
 
 user_data = get_user_data()
 amiId = get_ami_id()
+filename = '/var/ec2/.already-ran.%s' % amiId
 
-if user_data.startswith('#!'):
-    # run it 
-    (fp, path) = tempfile.mkstemp()
-    os.write(fp,user_data)
-    os.close(fp);
-    os.chmod(path, 0700)
-    os.system('cp %s /var/ec2/user-data.%s' %(path, amiId))
-    status = os.system('%s' % path)
-    os.unlink(path)
+if os.path.exists(filename):
+   print "ec2-run-user-data already ran for this instance."
+   sys.exit(0)
+else:
+    if user_data.startswith('#!'):
+       # run it 
+       (fp, path) = tempfile.mkstemp()
+       os.write(fp,user_data)
+       os.close(fp);
+       os.chmod(path, 0700)
+       os.system('cp %s /var/ec2/user-data' %(path))
+       status = os.system('%s' % path)
+       os.unlink(path)
+       os.system('touch /var/ec2/$s' %(filename))
 
 sys.exit(0)
